src/mppi_isaac/pyproject.toml
[tool.poetry]
name = "mppiisaac"
version = "0.1.0"
description = ""
authors = [
  "Corrado Pezzato <c.pezzato@tudelft.nl>",
  "Chadi Salmi <c.salmi@tudelft.nl",
  "Elia Trevisan <e.trevisan@tudelft.nl>",
  "Max Spahn <m.spahn@tudelft.nl>",
]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.8,<3.9"
pytest = "^7.2.1"
isaacgym = {path = "thirdparty/isaacgym/python", develop = true}
hydra-core = "^1.3.1"
pytorch3d = "^0.3.0"
zerorpc = "^0.6.3"
mppi-torch = {git = "ssh://git@github.com/tud-airlab/mppi_torch.git#75e17e8"}
tensorboard = "<2.12"

[tool.poetry.group.dev]
optional = true

[tool.poetry.group.dev.dependencies]
pylint = "^2.16.1"
black = "^23.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"